Move Base Node states to submodule
