+++ base_layer/core/src/base_node/states/initial_sync.rs
@@ -19,12 +19,10 @@
 // SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
 
 use crate::{
-    base_node::{
-        node_state::{StateEvent, StateEvent::FatalError},
-        starting_state::Starting,
-    },
+    base_node::states::{Starting, StateEvent, StateEvent::FatalError},
     chain_storage::{BlockchainBackend, BlockchainDatabase},
 };
 use log::*;
