+++ base_layer/core/src/base_node/states/starting_state.rs
@@ -19,8 +19,9 @@
 // SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
 use crate::{
-    base_node::{node_state::StateEvent, BaseNodeConfig},
+    base_node::{states::StateEvent, BaseNodeConfig},
     chain_storage::{BlockchainBackend, BlockchainDatabase},
 };
 use log::*;
